requirements.txt
streamlit>=1.36.0
pandas>=2.2.2
python-dotenv>=1.0.1
pydantic>=2.7.1
typing-extensions>=4.11.0
langchain>=0.2.16
langgraph>=0.2.41
langchain-core>=0.2.38
langchain-community>=0.2.17
langchain-nebius>=0.0.6
datasets>=2.20.0
rapidfuzz>=3.9.7
langgraph-checkpoint-sqlite>=0.0.8
aiosqlite>=0.20.0
